requirements.txt
kafka-python==2.0.2
ffmpeg-python==0.2.0
boto3==1.28.0
python-dotenv==1.0.0